fix:converted to scss and fsd
