+++ src/shared/lib/BookCardData.ts
