+++ src/widgets/book-card/index.ts
@@ -1,0 +1,1 @@
+export { BookCard } from './ui/BookCard'
