+++ src/widgets/book-card/ui/BookCard.module.scss
@@ -1,0 +1,111 @@
+@use '@/shared/styles/variables' as *;
+
+.card {
+  width: 242px;
+  height: 460px;
+  display: flex;
+  flex-direction: column;
+  border-radius: 16px;
+  cursor: pointer;
+  transition: box-shadow 0.2s ease;
+  background-color: $background-primary;
+
+  &:hover {
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
+  }
+}
+
+.imageWrapper {
+  width: 202px;
+  height: 304px;
+  margin: 20px 20px 0 20px;
+  position: relative;
+}
+
+.imageContainer {
+  position: relative;
+  width: 100%;
+  height: 100%;
+}
+
+.bookImage {
+  border-radius: 4px;
+  object-fit: cover;
+}
+
+.favoriteButton {
+  position: absolute;
+  top: 8px;
+  right: 8px;
+  background-color: rgba(255, 255, 255, 0.8);
+  border: none;
+  border-radius: 50%;
+  width: 28px;
+  height: 28px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+  padding: 4px;
+  z-index: 1;
+  transition: background-color 0.2s ease;
+
+  &:hover {
+    background-color: rgba(255, 255, 255, 0.9);
+  }
+
+  &:focus {
+    outline: none;
+  }
+}
+
+.favoriteIcon {
+  color: #FF3B3B;
+  font-size: 20px;
+}
+
+.favoriteBorderIcon {
+  color: #666;
+  font-size: 20px;
+}
+
+.content {
+  padding: 0 20px 20px 20px;
+}
+
+.title {
+  font-family: var(--font-literata), serif;
+  font-size: 14px;
+  font-weight: 500;
+  line-height: 130%;
+  color: $text-heading;
+  margin: 0 0 8px 0;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.author {
+  font-family: var(--font-onest), sans-serif;
+  font-size: 14px;
+  font-weight: 500;
+  line-height: 100%;
+  color: $color-primary;
+  margin: 0 0 12px 0;
+}
+
+.subtitle {
+  font-family: var(--font-onest), sans-serif;
+  font-size: 14px;
+  font-weight: 400;
+  line-height: 100%;
+  color: $text-secondary;
+  margin: 0;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
